+++ src/main/java/com/example/movieticketbookingbe/dto/systemsetting/SystemSettingDTO.java
@@ -18,7 +18,7 @@ public class SystemSettingDTO {
     private BigDecimal cancelFee;
     private Integer cancelTimeLimit;
 
-    private Integer priceSeatRegular;
-    private Integer priceSeatVip;
-    private Integer priceSeatDouble;
+    private Double  priceSeatRegular;
+    private Double  priceSeatVip;
+    private Double  priceSeatDouble;
 }
